EmployeeService, UsuarioService, EmployeeController e UsuarioController

+++ java-avancado/aula-10/sales-company/src/main/java/com/swproject/salescompany/services/UsuarioService.java
@@ -75,7 +75,11 @@ public class UsuarioService {
   }
 
   public void delete(@NonNull Long id) {
-    usuarioRepository.deleteById(id);
+    if(usuarioRepository.existsById(id)) {
+      usuarioRepository.deleteById(id);
+    } else {
+      throw new EntityNotFoundException(String.format("Usuário id=%s não encontrado", id));
+    }
   }
 
   public Boolean isExisteId(@NonNull Long id) {
